+++ src/app/features/focus-mode/focus-mode-task-selector/focus-mode-task-selector.component.ts
@@ -44,6 +44,7 @@ export class FocusModeTaskSelectorComponent implements AfterViewInit {
       this.inputElement =
         this.selectTaskElementRef?.nativeElement?.querySelector('input');
       this.inputElement?.focus();
+      this.selectTaskComponent?.openPanel();
     });
   }